fix: 修复抓拍链路 — YOLO float16 + 摄像头重连
- yolo_detector: ONNX模型需要float16输入，session.run()前加blob.astype(np.float16)
- simple_camera: 重写摄像头驱动，open时验证read()+自动重连机制
- 修复容器内摄像头isOpened()=True但read()失败的问题

+++ patrol_robot/patrol_robot/src/patrol_yolo/patrol_yolo/yolo_detector.py
@@ -168,12 +168,14 @@ class YoloDetector(Node):
 
         # 1. 预处理: resize + normalize
         blob = cv2.resize(img, (self.input_width, self.input_height))
-        blob = blob.astype(np.float32) / 255.0
+        blob = blob.astype(np.float16) / 255.0
         blob = np.transpose(blob, (2, 0, 1))  # HWC → CHW
         blob = np.expand_dims(blob, axis=0)    # → (1, 3, 640, 640)
 
         # 2. ONNX 推理
         with self._lock:
+            blob = blob.astype(np.float16)
+            blob = blob.astype(np.float16)
             outputs = self.session.run(None, {self.input_name: blob})
 
         # YOLOv5 输出: (1, 25200, 85) — 85 = 4(bbox) + 1(obj) + 80(cls)
